# swagger.yaml
openapi: 3.0.3
info:
  title: Task / Issue Management
  description: Simple Task/Issue Management API build on NodeJS
  version: 0.0.1
servers:
  - url: http://localhost:3000
    description: local testing erver
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
...
paths:
...
  /task:
    get:
      security:
        - baseAuth:
      tags:
        - Task
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: Success
        400:
          description: Bad request
        500:
          description: Internal server exception
        401:
          description: Unauthorized
    post:
      security:
        - bearerAuth:
      tags:
        - Task
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        x-name: body
        content:
          application/json:
            schema:
              $ref: "#components/schemas/create-task-input"
      responses:
        201:
          description: Success
        400:
          description: Bad request
        500:
          description: Internal server exception
        401:
          description: Unauthorized
  /task/{id}:
    get:
      security:
        - bearerAuth:
      tags:
        - Task
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: user's UUID
      responses:
        200:
          description: Success
        400:
          description: Bad request
        500:
          description: Internal server exception
        401:
          description: Unauthorized
    patch:
      security:
        - bearerAuth:
      tags:
        - Task
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: user's UUID
      requestBody:
        x-name: body
        content:
          application/json:
            schema:
              $ref: "#components/schemas/update-task-input"
      responses:
        203:
          description: Success
        400:
          description: Bad request
        500:
          description: Internal server exception
        401:
          description: Unauthorized

  # user routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, for documentation purposes only
...